# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Telecom Management - API Documentation",
        "version": "1.0"
    },
    "paths": {
...
        "/api/cucmcdrs/search/{column}/{search}": {
            "get": {
                "tags": [
                    "CDR - CUCM History"
                ],
                "summary": "List CDR Records by Date Range",
                "description": "",
                "operationId": "searchCDR",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "column",
                        "in": "path",
                        "description": "Column to Search",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "search",
                        "in": "path",
                        "description": "Search",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
        "/api/sonus5kcdrs/search/{column}/{search}": {
            "get": {
                "tags": [
                    "CDR - Sonus History"
                ],
                "summary": "List CDR Records by Date Range",
                "description": "",
                "operationId": "searchCDR",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "column",
                        "in": "path",
                        "description": "Column to Search",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "search",
                        "in": "path",
                        "description": "Search",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    },
...
}